# api/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Controller API
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: http://localhost:8081
  - url: http://localhost:8082
  - url: http://localhost:8888
paths:
...
  /gateway/document/{statementId}/send:
    post:
      summary: Отправка запроса на получение документов
      tags:
        - Gateway
      operationId: sendRequestForDocument
      parameters:
        - in: path
          name: statementId
          schema:
            type: string
          required: true
          description: Идентификатор заявки
      responses:
        '204':
          description: No Content
        '400':
          description: Неверный идентификатор заявки
        '404':
          description: Заявка не найдена
        '500':
          description: Внутренняя ошибка сервера
...